// dto/ssl.dto.ts
import { CreateFileDTO, EditFileDTO } from "./files.dto";

export interface SSLFormUpdateDto {
  id: number;
  createdById?: number;
  assignedToId?: number;
  status?: string;
  address1?: string;
  address2?: string;
  address3?: string;
  address4?: string;
  city?: string;
  country?: string;
  postCode?: string;
  noAcPoints?: number;
  noDcPoints?: number;
  makeOfChargePoint?: string;
  ampsPerCp?: string;
  kwPerCp?: string;
  wallMountSingle?: number;
  wallMountDual?: number;
  floorMountSingle?: number;
  floorMountDual?: number;
  phaseType?: string;
  mainsFuseSize?: string;
  mainsIsolation?: string;
  earthingSetup?: string;
  mobileSignal?: string;
  signalStrength?: string;
  consumerUnitMake?: string;
  consumerUnitModel?: string;
  totalSpareWays?: string;
  totalAmpsInUse?: string;
  cableSizeRead?: string;
  cableLengthTotal?: string;
  totalCableRuns?: string;
  internalCableAttachment?: string;
  dataCableRun?: string;
  cableRunDescription?: string;
  selectOptions?: string;
  gwLength?: string;
  gwWidth?: string;
  gwDepth?: string;
  ductingSize?: string;
  ductingLength?: string;
  groundWorksDescription?: string;
  installationEarthingSetup?: string;
  files: CreateFileDTO[];
}

export interface CreateSSLFormDto {
  [key: string]: any;
  createdById: number;
  assignedToId: number;
  status: string;
  address1: string;
  address2?: string;
  address3?: string;
  address4?: string;
  city: string;
  country: string;
  postCode: string;
  noAcPoints: number;
  noDcPoints: number;
  makeOfChargePoint: string;
  ampsPerCp: string;
  kwPerCp: string;
  wallMountSingle: number;
  wallMountDual: number;
  floorMountSingle: number;
  floorMountDual: number;
  phaseType: string;
  mainsFuseSize: string;
  mainsIsolation: string;
  earthingSetup: string;
  mobileSignal: string;
  signalStrength: string;
  consumerUnitMake: string;
  consumerUnitModel: string;
  totalSpareWays: string;
  totalAmpsInUse: string;
  cableSizeRead: string;
  cableLengthTotal: string;
  totalCableRuns: string;
  internalCableAttachment: string;
  dataCableRun: string;
  cableRunDescription: string;
  selectOptions: string;
  gwLength: string;
  gwWidth: string;
  gwDepth: string;
  ductingSize: string;
  ductingLength: string;
  groundWorksDescription: string;
  installationEarthingSetup: string;
  files: CreateFileDTO[];
}